Added CCTVFocusManager and related classes.
I've added the Cocos2D 3.x version of my CCTVMenu class to the repo for others to use.

I should note that although I believe it works, I've not shipped an app that uses it.  There was also an added complexity with the use of CCControl subclasses in that they are typically children of the scene, and not of a "menu" object as in Cocos2D 2.x.  For me, the solution I came up with was to have proxy objects that wrapped the CCControl objects and those proxies are then added to the focus manager.  It's not as nice as the v2.2 arrangement, but it does work whilst minimising the amount of existing code that needed changing, especially in an app where the code is being shared between iOS and tvOS.

I also note that I've provided an attempt to make CCTextField focusable however I hit serious problems because Apple's own focus manager would interfere.  My solution was to make the text field invisible when it doesn't have focus, and to display a label in it's place.  This mechanism is mentioned in comments, but not completely finished within CCFocusableTextField (I got a prototype working within my app).

The basic mechanisms provided do work, and work on quite complex scenes with lot's of controls, but it does need tuning.  It's certainly not as polished as Apple's focus manager.